requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt ./pyproject.toml
#
aio-pika==9.3.1
    # via omotes-sdk-python
aiormq==6.7.7
    # via aio-pika
amqp==5.2.0
    # via kombu
apispec[marshmallow]==6.3.0
    # via flask-smorest
billiard==4.2.0
    # via celery
blinker==1.6.2
    # via flask
celery==5.3.6
    # via omotes-sdk-python
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   flask
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via
    #   click
    #   omotes-rest (pyproject.toml)
flask==2.3.3
    # via
    #   flask-cors
    #   flask-dotenv
    #   flask-smorest
    #   omotes-rest (pyproject.toml)
flask-cors==4.0.0
    # via omotes-rest (pyproject.toml)
flask-dotenv==0.1.2
    # via omotes-rest (pyproject.toml)
flask-smorest==0.42.1
    # via omotes-rest (pyproject.toml)
greenlet==2.0.2
    # via sqlalchemy
gunicorn==21.2.0
    # via omotes-rest (pyproject.toml)
idna==3.6
    # via yarl
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
kombu==5.3.5
    # via celery
mako==1.3.2
    # via pdoc3
markdown==3.6
    # via pdoc3
markupsafe==2.1.3
    # via
    #   jinja2
    #   mako
    #   werkzeug
marshmallow==3.20.1
    # via
    #   apispec
    #   flask-smorest
    #   marshmallow-dataclass
    #   marshmallow-enum
    #   omotes-rest (pyproject.toml)
    #   webargs
marshmallow-dataclass==8.5.14
    # via omotes-rest (pyproject.toml)
marshmallow-enum==1.5.1
    # via omotes-rest (pyproject.toml)
multidict==6.0.5
    # via yarl
mypy==1.5.1
    # via sqlalchemy-stubs
mypy-extensions==1.0.0
    # via
    #   mypy
    #   typing-inspect
omotes-sdk-protocol==0.0.8
    # via omotes-sdk-python
omotes-sdk-python==0.0.12
    # via omotes-rest (pyproject.toml)
packaging==23.1
    # via
    #   apispec
    #   gunicorn
    #   marshmallow
    #   webargs
pamqp==3.2.1
    # via aiormq
pdoc3==0.10.0
    # via streamcapture
prompt-toolkit==3.0.43
    # via click-repl
protobuf==4.25.3
    # via omotes-sdk-protocol
python-dateutil==2.9.0.post0
    # via celery
python-dotenv==1.0.0
    # via omotes-rest (pyproject.toml)
six==1.16.0
    # via python-dateutil
sqlalchemy==2.0.28
    # via omotes-rest (pyproject.toml)
sqlalchemy-stubs==0.4
    # via omotes-rest (pyproject.toml)
streamcapture==1.2.2
    # via omotes-sdk-python
structlog==23.1.0
    # via omotes-rest (pyproject.toml)
tomli==2.0.1
    # via mypy
typing-extensions==4.7.1
    # via
    #   marshmallow-dataclass
    #   mypy
    #   sqlalchemy
    #   sqlalchemy-stubs
    #   typing-inspect
typing-inspect==0.9.0
    # via marshmallow-dataclass
tzdata==2024.1
    # via celery
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
webargs==8.3.0
    # via flask-smorest
werkzeug==2.3.7
    # via
    #   flask
    #   flask-smorest
yarl==1.9.4
    # via
    #   aio-pika
    #   aiormq

# The following packages are considered to be unsafe in a requirements file:
# setuptools
